requirements.txt
python-dotenv
fastapi
openai